requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
